requirements.txt
google-api-python-client==2.175.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
hubspot-api-client==12.0.0
langchain-groq==0.3.2
langchain==0.3.24
